{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "cookiecutter-django"
version = "0.1.0"
description = ""
authors = ["Miha Zelnik <miha.zelnik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pytz = "^2019.2"
python-slugify = "^3.0"
Pillow = "^6.1"
argon2-cffi = "^19.1"
whitenoise = "^4.1"
redis = "^3.3"
django = "^2.2"
django-environ = "^0.4.5"
django-model-utils = "^3.2"
django-redis = "^4.10"
djangorestframework = "^3.10"
psycopg2 = "^2.8"
gunicorn = "^19.9"
sentry-sdk = "^0.11.1"
django-anymail = "^6.1"
drf-yasg = "^1.16"
django-cors-headers = "^3.1"
djangorestframework_simplejwt = "^4.3"
django-timezone-field = "^3.0"

[tool.poetry.dev-dependencies]
Werkzeug = "^0.15.5"
ipdb = "^0.12.2"
pytest = "^5.1"
pytest-sugar = "^0.9.2"
flake8 = "^3.7"
pylint-django = "^2.0"
django-debug-toolbar = "^2.0"
django-extensions = "^2.2"
pytest-django = "^3.5"
black = {version = "*",allows-prereleases = true}
factory-boy = "^2.12"
isort = "^4.3"

[tool.black]
line-length = 100
target-version = ['py37', 'py38']
skip-numeric-underscore-normalization = true
include = '\.py$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

